+++ full-application/src/main/java/com/evolutionnext/application/commands/customer/CreateCustomer.java
@@ -3,5 +3,7 @@ package com.evolutionnext.application.commands.customer;
 
 import com.evolutionnext.domain.aggregates.customer.CustomerId;
 
-public record CreateCustomer(CustomerId customerId, String name, float creditLimit) implements CustomerCommand {
+import java.math.BigDecimal;
+
+public record CreateCustomer(CustomerId customerId, String name, BigDecimal creditLimit) implements CustomerCommand {
 }
