+++ full-application/src/main/java/com/evolutionnext/application/commands/order/AddOrderItem.java
@@ -4,5 +4,7 @@ package com.evolutionnext.application.commands.order;
 import com.evolutionnext.domain.aggregates.order.OrderId;
 import com.evolutionnext.domain.aggregates.product.ProductId;
 
-public record AddOrderItem(OrderId orderId, ProductId productId, int quantity, int price) implements ClientOrderCommand {
+import java.math.BigDecimal;
+
+public record AddOrderItem(OrderId orderId, ProductId productId, int quantity, BigDecimal price) implements ClientOrderCommand {
 }
