+++ full-application/src/main/java/com/evolutionnext/domain/aggregates/customer/Customer.java
@@ -1,5 +1,7 @@
 package com.evolutionnext.domain.aggregates.customer;
 
 
-public record Customer(CustomerId id, String name, double creditLimit) {
+import java.math.BigDecimal;
+
+public record Customer(CustomerId id, String name, BigDecimal creditLimit) {
 }
