+++ full-application/src/main/java/com/evolutionnext/domain/aggregates/order/OrderItem.java
@@ -3,5 +3,7 @@ package com.evolutionnext.domain.aggregates.order;
 
 import com.evolutionnext.domain.aggregates.product.ProductId;
 
+import java.math.BigDecimal;
+
 public record OrderItem(ProductId productId,
-                        int quantity, int price) {}
+                        int quantity, BigDecimal price) {}
