+++ full-application/src/main/java/com/evolutionnext/dto/order/OrderData.java
@@ -4,6 +4,8 @@ package com.evolutionnext.dto.order;
 import com.evolutionnext.domain.aggregates.customer.CustomerId;
 import com.evolutionnext.domain.aggregates.order.OrderId;
 
-public record OrderData(OrderId orderId, CustomerId customerId, String name, float total) {
+import java.math.BigDecimal;
+
+public record OrderData(OrderId orderId, CustomerId customerId, String name, BigDecimal total) {
 
 }
